+++ frb_codegen/src/library/codegen/parser/mir/parser/function/real/mod.rs
@@ -15,7 +15,7 @@ use crate::codegen::ir::mir::ty::trait_def::MirTypeTraitDef;
 use crate::codegen::ir::mir::ty::MirType;
 use crate::codegen::parser::mir::internal_config::ParserMirInternalConfig;
 use crate::codegen::parser::mir::parser::attribute::FrbAttributes;
-use crate::codegen::parser::mir::parser::function::real::structs::MirFuncOrSkip;
+use crate::codegen::parser::mir::parser::function::func_or_skip::MirFuncOrSkip;
 use crate::codegen::parser::mir::parser::ty::misc::parse_comments;
 use crate::codegen::parser::mir::parser::ty::trait_def::parse_type_trait;
 use crate::codegen::parser::mir::parser::ty::{TypeParser, TypeParserParsingContext};
@@ -31,7 +31,6 @@ use MirType::Primitive;
 
 pub(crate) mod argument;
 pub(crate) mod output;
-pub(crate) mod structs;
 mod transformer;
 
 pub(crate) fn parse(
